requirements.txt
Django==4.2.0
djangorestframework==3.14.0
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
drf-yasg==1.21.5
gunicorn==21.2.0
django-cors-headers==4.1.0
pytest==7.3.1
pytest-django==4.5.2
